% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/federation_search.R
\name{federation_search}
\alias{federation_search}
\title{Summary endpoint}
\usage{
federation_search(query, strict = FALSE, snippetCount = 0,
  expandSynonyms = TRUE, expandAcronyms = FALSE, expandAbbrevs = FALSE,
  expandInferred = TRUE, ...)
}
\arguments{
\item{query}{Query terms}

\item{strict}{(logical) Only consider more relevant columns. Default: \code{FALSE}}

\item{snippetCount}{How many snippets to return. Default: 0}

\item{expandSynonyms}{(logical) Should recognized terms in the query have synonyms
expanded? Default: \code{TRUE}}

\item{expandAcronyms}{(logical) Default: \code{FALSE}}

\item{expandAbbrevs}{(logical) Default: \code{FALSE}}

\item{expandInferred}{(logical) Should recognized terms in the query have inferred terms
expanded? Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Summary endpoint
}
\examples{
\dontrun{
out <- federation_search("cellular")
out$query
out$messages
out$result$literatureSummary
out$result$federationSummary$total
out$result$federationSummary$categories
head(out$result$federationSummary$results)

head( summary("cellular")$result$federationSummary$results )
head( summary("cellular", expandSynonyms=FALSE)$result$federationSummary$results )
head( summary("cellular", expandAbbrevs=TRUE)$result$federationSummary$results )

library("httr")
summary("cellular", config=verbose())
}
}

