% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literature_retractions.R
\name{literature_retractions}
\alias{literature_retractions}
\title{Search literature retractions}
\usage{
literature_retractions(key = NULL, ...)
}
\arguments{
\item{key}{(character) API key. required either passed in here or save
as an env var or R option. see \code{\link{rif}} and \code{\link{Startup}}}

\item{...}{Curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\description{
Search literature retractions
}
\examples{
\dontrun{
# get PMIDs for retracted articles
(out <- literature_retractions())
# then get info on some articles
literature_pmid(pmid = out[1])
}
}
\seealso{
Other literature: \code{\link{literature_mlpmid}},
  \code{\link{literature_mltext}},
  \code{\link{literature_pmid}}
}
