% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rif_query_categories.R
\name{rif_query_categories}
\alias{rif_query_categories}
\title{Query NIF data}
\usage{
rif_query_categories(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Query NIF data
}
\examples{
\dontrun{
rif_query_categories()

library("httr")
rif_query_categories(config = verbose())
}
}
\seealso{
Other query: \code{\link{rif_query}}
}

