% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_categories.R
\name{vocabulary_categories}
\alias{vocabulary_categories}
\title{Vocabulary services - get categories}
\usage{
vocabulary_categories(vocabulary = NULL, ...)
}
\arguments{
\item{vocabulary}{A vocabulary}

\item{...}{Curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\description{
Vocabulary services - get categories
}
\examples{
\dontrun{
vocabulary_categories()
}
}
\seealso{
Other vocabulary: \code{\link{vocabulary_providers}},
  \code{\link{vocabulary_search}},
  \code{\link{vocabulary_suggest}},
  \code{\link{vocabulary}}
}
