% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_providers.R
\name{vocabulary_providers}
\alias{vocabulary_providers}
\title{Vocabulary services - get providers}
\usage{
vocabulary_providers(vocabulary = NULL, key = NULL, ...)
}
\arguments{
\item{vocabulary}{A vocabulary}

\item{key}{(character) API key. required either passed in here or save
as an env var or R option. see \code{\link{rif}} and \code{\link{Startup}}}

\item{...}{Curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\description{
Vocabulary services - get providers
}
\examples{
\dontrun{
vocabulary_providers()
}
}
\seealso{
Other vocabulary: \code{\link{vocabulary_categories}},
  \code{\link{vocabulary_search}},
  \code{\link{vocabulary_suggest}},
  \code{\link{vocabulary}}
}
