% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_search.R
\name{vocabulary_search}
\alias{vocabulary_search}
\title{Vocabulary services - Search for vocabulary terms and their metadata}
\usage{
vocabulary_search(term, limit = 20, category = NULL, vocabulary = NULL,
  key = NULL, ...)
}
\arguments{
\item{term}{(character) The search terms}

\item{limit}{The number of candidates to return. Default: 20}

\item{category}{A category}

\item{vocabulary}{A vocabulary}

\item{key}{(character) API key. required either passed in here or save
as an env var or R option. see \code{\link{rif}} and \code{\link{Startup}}}

\item{...}{Curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\description{
Vocabulary services - Search for vocabulary terms and their metadata
}
\examples{
\dontrun{
vocabulary_search("cell")
vocabulary_search("cell", limit = 3)
vocabulary_search("cell", category = "biological_process")
}
}
\seealso{
Other vocabulary: \code{\link{vocabulary_categories}},
  \code{\link{vocabulary_providers}},
  \code{\link{vocabulary_suggest}},
  \code{\link{vocabulary}}
}
