% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_suggest.R
\name{vocabulary_suggest}
\alias{vocabulary_suggest}
\title{Vocabulary services - Get suggestions for terms}
\usage{
vocabulary_suggest(query, limit = 10, ...)
}
\arguments{
\item{query}{(character) Query terms}

\item{limit}{The number of candidates to return. Default: 20}

\item{...}{Curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\description{
DOESN'T SEEM TO WORK
}
\examples{
\dontrun{
vocabulary_suggest("cell")
}
}
\seealso{
Other vocabulary: \code{\link{vocabulary_categories}},
  \code{\link{vocabulary_providers}},
  \code{\link{vocabulary_search}}, \code{\link{vocabulary}}
}
