% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literature_mltext.R
\name{literature_mltext}
\alias{literature_mltext}
\title{Literature - more like text search}
\usage{
literature_mltext(text, offset = 0, count = 20, ...)
}
\arguments{
\item{text}{The query}

\item{offset}{(integer) The result (row) to start on    query    0}

\item{count}{(integer) The number of results (rows) to return. Default: 5}

\item{...}{Curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\description{
Literature - more like text search
}
\examples{
\dontrun{
literature_mltext(text = "cellular")
literature_mltext(text = "ecology", count = 200)
}
}
\seealso{
Other literature: \code{\link{literature_mlpmid}},
  \code{\link{literature_pmid}},
  \code{\link{literature_retractions}}
}
