% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rif-package.R
\docType{package}
\name{rif-package}
\alias{rif}
\alias{rif-package}
\title{R Client for Neuroscience Information Framework APIs}
\description{
R Client for Neuroscience Information Framework APIs
}
\keyword{package}

