% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rif-package.R
\docType{package}
\name{rif-package}
\alias{rif-package}
\alias{rif}
\title{R Client for Neuroscience Information Framework APIs}
\description{
R Client for Neuroscience Information Framework APIs
}
\note{
The following services are not available in this client:
\itemize{
\item annotate - It appears that this is mostly for browser based
annotation, so not a good fit for an R package
}
}
\author{
\email{myrmecocystus@gmail.com}
}
\keyword{package}
