% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rif_query_categories.R
\name{rif_query_categories}
\alias{rif_query_categories}
\title{Query NIF data}
\usage{
rif_query_categories(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{query}{Query terms}

\item{expandSynonyms}{(logical) Should recognized terms in the query have synonyms
expanded? Default: \code{TRUE}}

\item{expandAcronyms}{(logical) Default: \code{FALSE}}

\item{expandAbbrevs}{(logical) Default: \code{FALSE}}

\item{expandInferred}{(logical) Should recognized terms in the query have inferred terms
expanded? Default: \code{TRUE}}
}
\description{
Query NIF data
}
\examples{
\dontrun{
rif_query_categories()

library("httr")
rif_query_categories(config = verbose())
}
}
\seealso{
Other query: \code{\link{rif_query}}
}

